removed a condition that was causing hearts to be removed. and corrected developer error ex ==, ===
the if statement in my qwerty.listener was checking if the target had the class 'chosen'  and returning null because of it. this was causing my if statement on line 70  to be true when I did not intend it to be. Therfore wrongly removing a childElement and  adding +1  to the missedGuesses. Now buttons can be clicked more than once for words than contain multiples of a letter example 'books'.